requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
openai
requests
pydantic
motor